Move Orama search index to a dedicated web worker
iOS Safari kills the tab when the main thread spikes memory during Orama
index build/deserialisation. Running all index operations in a Worker
isolates that memory from the page heap so the browser can manage it
independently without crashing the tab.

- searchIndexCore.ts: extracted Orama logic (schema, index build,
  cache, search helpers) — used directly by the worker and as a
  fallback in environments where Worker is unavailable (e.g. tests).
- searchIndex.worker.ts: thin onmessage wrapper around the core
  functions; all Orama state lives in the worker's heap.
- searchIndex.ts: worker client with message-based Promise API;
  falls back to the core directly when Worker is not available so
  existing tests continue to run without changes.
- Reduces explorer search tolerance from 2 to 1 to ease fuzzy-match
  CPU/memory pressure on constrained devices.
